Add base class and inherit tests for V1 E2E
add tests from v1 E2E, remove old E2E class
